fixed unsynchronized visited link db insertion
When doing FinTen.fill, the visited link insertion to the DB would
go out of sync if more than one downloadable was downloaded. It
would keep adding the same visited link over and overt again.
The modifications made in the DownloadManager class solve this problem,
as the returned value now is an array of Downloadables (see interface
for more details).