fixed XBRL parsing problems
Some documents had more than one XBRL tag, and we were only reading
the first one. Sometimes that XBRL was not parsable, but one of the
subsequents ones was. The code has been modified to retrieve all the
available XBRLs in the given document and return only the one that
can be parsed.

Test cases have been added to ensure this mechanism works.