fix: fixed default logging issues
DefaultLogger was configured to log in './logs/.log'. The 'logs' folder
does not exist in the repository by default, so when deploying,
initilization would fail because the WriteStream could not be created
due to ENOENT error (file doesn't exist).
Adding the folder to the repo and modifying the way the WriteStream is
created (in 'as+', open for appending in synchronous mode, and create it
if it doesn't exist) should fix the issue.