fix: test cases did not have the proper db setup
Only the first suite in the tests would have it properly, which caused
all other tests to fail when that one was not executed. The problem was
that reconnecting FinTenDB requires for it to be firstly disconnected,
which was only happening in the first test suite.
All test cases that access FinTenDB have now been modified to not depend
on the execution of that first test suite.