fix: some tests no longer use network
Fred tests now check if the proper url is hit by checking the series_id
parameter of the url. They do not use the network anymore.

Also, the learning tests to understand the Fred api have been moved to
a file of type *.spec.learning.ts, which should be from now on the
convetion for all learning tests. A dedicated npm script has been added
to the package.json to exclusively test these.