feat: demo filings retrieved from any demo ticker
Also fixed the problem when setting the logger to save the logs in
a file the path of which did not exist. The whole application would stop
due to 'file does not exist' error. The solution was to basically create
the whole path and the file in case it didn't exist.