refactor: queues with generics and single '.get'
The Queue interface has been modified to use generics instead of having
Downloadable hardcoded in it. The DownloadManager still has it
hardcoded, maybe that can be fixed in the future.

Also, implemented a single '.get()' in DownloadManager to be able to
download only one Downloadable and get a single element back. It
happened that in FinTen, when adding new filings, the get method was
called with only one argument, and an array with one element was
returned, which requires us to loop through the array. In order to make
the code less cumbersome (and because we don't need a varargs '.get()'),
an effort was made to modify the method in such a way.