feat: returned filings have the requested ticker
Instead return the ticker that the parser and our post-parsing
processing add to the filing, return the one the user asked for to avoid
any sort of confusion about this. In the end, the ticker doesn't really
matter when it comes to filings because it's the same entity that filed
them.