fix: demo filings
Returning the first 4 known filings. Otherwise, people could pay for the
first time, get all old filings, and then rely on the demo endpoints to
get newer filings.
Also, the ticker of the filings is reassigned to be equal to that of the
request, just like in the normal endpoint.