fix: check if macro observation was already known
When saving macro observation we were relying on the error message of
Mongoose to know if it was duplicate or not and avoid logging that
error. But error messages can change.
Instead, we now first check if there is a document in the corresponding
collection with the same date as the observation, and in that case, just
ignore that observation.